Initial commit with qvw file and analysis report
This commit has 4 files including the .qvw file which has the data model as well as the various dashboards and analysis. The Final script is the file having the script for creating the data model. The ppt attached has the report and the summary doc has pictures of the dashboard and the filters of reports and their insights.